+++ declarations/android/android-platform-35.d.ts
@@ -168543,4 +168543,3 @@ declare module org {
 //java.util.stream.Collector:3
 //java.util.stream.Stream:1
 //java.util.stream.Stream.Builder:1
-